requirements.txt
python-telegram-bot==13.15
pandas==2.0.3
numpy==1.24.3
yfinance==0.2.31
python-dotenv==1.0.0
APScheduler==3.6.3
flask==2.3.3